Move render_sorting_header template tag to dashboard
