+++ templates/dashboard/category/detail.html
@@ -2,7 +2,6 @@
 {% load i18n %}
 {% load mptt_tags %}
 {% load materializecss %}
-{% load render_sorting_header from shop %}
 {% load staticfiles %}
 {% load utils %}
