+++ templates/dashboard/discount/sale/list.html
@@ -1,7 +1,6 @@
 {% extends "dashboard/base.html" %}
 {% load i18n %}
 {% load prices_i18n %}
-{% load render_sorting_header from shop %}
 {% load staticfiles %}
 {% load materializecss %}
 {% load utils %}
