+++ templates/dashboard/group/list.html
@@ -1,7 +1,6 @@
 {% extends "dashboard/base.html" %}
 {% load i18n %}
 {% load materializecss %}
-{% load render_sorting_header from shop %}
 {% load staticfiles %}
 {% load utils %}
