+++ templates/dashboard/includes/_orders_table.html
@@ -1,6 +1,5 @@
 {% load i18n %}
 {% load prices_i18n %}
-{% load render_sorting_header from shop %}
 {% load status %}
 
 <div class="data-table-container">
