+++ templates/dashboard/product/list.html
@@ -2,7 +2,6 @@
 {% load i18n %}
 {% load prices_i18n %}
 {% load materializecss %}
-{% load render_sorting_header from shop %}
 {% load staticfiles %}
 {% load product_first_image from product_images %}
 {% load utils %}
