+++ templates/dashboard/product/product_class/list.html
@@ -2,7 +2,6 @@
 {% load i18n %}
 {% load prices_i18n %}
 {% load materializecss %}
-{% load render_sorting_header from shop %}
 {% load staticfiles %}
 {% load utils %}
